requirements.txt
python-dotenv
pdfplumber
openai
litellm
google-adk
google-agent-development-kit @ git+https://github.com/google/agent-development-kit.git
